+++ sparv/core/io.py
@@ -3,15 +3,16 @@
 import bz2
 import gzip
 import heapq
+import logging
 import lzma
 import os
-import re
+import pickle
 from pathlib import Path
-from typing import List, Optional, Tuple, Union
+from typing import Any, Iterator, List, Optional, Tuple, Union
 
 from sparv.api.classes import BaseAnnotation, BaseOutput
 from sparv.core import paths
-from sparv.core.misc import get_logger, SparvErrorMessage
+from sparv.core.misc import SparvErrorMessage, get_logger
 
 logger = get_logger(__name__)
 
@@ -34,20 +35,23 @@ _compressed_open = {
 }
 
 
-def annotation_exists(annotation: BaseAnnotation, source_file: Optional[str] = None):
+def annotation_exists(annotation: BaseAnnotation, source_file: Optional[str] = None) -> bool:
     """Check if an annotation file exists."""
     annotation_path = get_annotation_path(source_file or annotation.source_file, annotation, data=annotation.data)
-    return os.path.exists(annotation_path)
+    return annotation_path.exists()
 
 
-def remove_annotation(annotation: BaseAnnotation, source_file: Optional[str] = None):
+def remove_annotation(annotation: BaseAnnotation, source_file: Optional[str] = None) -> None:
     """Remove an annotation file."""
     annotation_path = get_annotation_path(source_file or annotation.source_file, annotation, data=annotation.data)
     annotation_path.unlink(missing_ok=True)
 
 
-def write_annotation(source_file: str, annotation: BaseOutput, values, append: bool = False,
-                     allow_newlines: bool = False) -> None:
+def write_annotation(
+    source_file: str,
+    annotation: BaseOutput,
+    values: list
+) -> None:
     """Write an annotation to one or more files. The file is overwritten if it exists.
 
     The annotation should be a list of values.
@@ -56,7 +60,7 @@ def write_annotation(source_file: str, annotation: BaseOutput, values, append: b
 
     if len(annotations) == 1:
         # Handle single annotation
-        _write_single_annotation(source_file, annotations[0], values, append, annotation.root, allow_newlines)
+        _write_single_annotation(source_file, annotations[0], values, annotation.root)
     else:
         elem_attrs = dict(split_annotation(ann) for ann in annotations)
         # Handle multiple annotations used as one
@@ -65,83 +69,56 @@ def write_annotation(source_file: str, annotation: BaseOutput, values, append: b
         # Get spans and associated names for annotations. We need this information to figure out which value goes to
         # which annotation.
         spans = read_annotation(source_file, annotation, with_annotation_name=True, spans=True)
-        annotation_values = {elem: [] for elem in elem_attrs.keys()}
+        annotation_values = {elem: [] for elem in elem_attrs}
 
         for value, (_, annotation_name) in zip(values, spans):
             annotation_values[annotation_name].append(value)
 
         for annotation_name in annotation_values:
             _write_single_annotation(source_file, join_annotation(annotation_name, elem_attrs[annotation_name]),
-                                     annotation_values[annotation_name], append, annotation.root, allow_newlines)
+                                     annotation_values[annotation_name], annotation.root)
 
 
-def _write_single_annotation(source_file: str, annotation: str, values, append: bool, root: Path,
-                             allow_newlines: bool = False):
+def _write_single_annotation(
+    source_file: str,
+    annotation: str,
+    values: list,
+    root: Path
+) -> None:
     """Write an annotation to a file."""
     is_span = not split_annotation(annotation)[1]
 
     if is_span:
         # Make sure that spans are sorted
         assert all(values[i] <= values[i + 1] for i in range(len(values) - 1)), "Annotation spans must be sorted."
+        # Always save spans with decimal tuples
+        if values and not isinstance(values[0][0], tuple):
+            values = [((v[0],), (v[1],)) for v in values]
     file_path = get_annotation_path(source_file, annotation, root)
-    os.makedirs(os.path.dirname(file_path), exist_ok=True)
-    mode = "a" if append else "w"
-    with open_annotation_file(file_path, mode) as f:
-        ctr = 0
-        for value in values:
-            if value is None:
-                value = ""
-            elif is_span:
-                start, end = value
-                start_subpos, end_subpos = None, None
-                if isinstance(start, tuple):
-                    start, start_subpos = start
-                if isinstance(end, tuple):
-                    end, end_subpos = end
-                start_subpos = ".{}".format(start_subpos) if start_subpos is not None else ""
-                end_subpos = ".{}".format(end_subpos) if end_subpos is not None else ""
-                value = "{}{}-{}{}".format(start, start_subpos, end, end_subpos)
-            elif allow_newlines:
-                # Replace line breaks with "\n"
-                value = value.replace("\\", r"\\").replace("\n", r"\n").replace("\r", "")
-            else:
-                # Remove line breaks entirely
-                value = value.replace("\n", "").replace("\r", "")
-            print(value, file=f)
-            ctr += 1
+    file_path.parent.mkdir(parents=True, exist_ok=True)
+    write_annotation_file(file_path, values)
     # Update file modification time even if nothing was written
     os.utime(file_path, None)
-    logger.info("Wrote %d items: %s%s", ctr, source_file + "/" if source_file else "", annotation)
+    logger.info("Wrote %d items: %s%s%s", len(values), source_file, "/" if source_file else "", annotation)
 
 
-def get_annotation_size(source_file: str, annotation: BaseAnnotation):
+def get_annotation_size(source_file: str, annotation: BaseAnnotation) -> int:
     """Return number of lines in an annotation."""
-    def _generator(reader_):
-        while True:
-            b = reader_(2 ** 16)
-            if not b:
-                break
-            yield b
-
     count = 0
 
     for ann in annotation.name.split():
         ann_file = get_annotation_path(source_file, ann, annotation.root)
-
-        try:
-            with open_annotation_file(ann_file, mode="rb") as f:
-                reader = f.raw.read if hasattr(f, "raw") and hasattr(f.raw, "read") else f.read
-                count += sum(buf.count(b"\n") for buf in _generator(reader))
-        except (gzip.BadGzipFile, OSError, lzma.LZMAError, UnicodeDecodeError) as e:
-            if isinstance(e, OSError) and str(e) != "Invalid data stream":
-                raise e
-            raise_format_error(ann_file)
+        count += len(list(read_annotation_file(ann_file)))
 
     return count
 
 
-def read_annotation_spans(source_file: str, annotation: BaseAnnotation, decimals: bool = False,
-                          with_annotation_name: bool = False):
+def read_annotation_spans(
+    source_file: str,
+    annotation: BaseAnnotation,
+    decimals: bool = False,
+    with_annotation_name: bool = False
+) -> Iterator[tuple]:
     """Iterate over the spans of an annotation."""
     # Strip any annotation attributes
     for span in read_annotation(source_file, annotation, with_annotation_name, spans=True):
@@ -151,8 +128,12 @@ def read_annotation_spans(source_file: str, annotation: BaseAnnotation, decimals
             yield span
 
 
-def read_annotation(source_file: str, annotation: BaseAnnotation, with_annotation_name: bool = False,
-                    allow_newlines: bool = False, spans: bool = False):
+def read_annotation(
+    source_file: str,
+    annotation: BaseAnnotation,
+    with_annotation_name: bool = False,
+    spans: bool = False
+) -> Iterator:
     """Yield each line from an annotation file."""
     if spans:
         annotations = [split_annotation(ann)[0] for ann in annotation.name.split()]
@@ -161,99 +142,86 @@ def read_annotation(source_file: str, annotation: BaseAnnotation, with_annotatio
     root = annotation.root
     if len(annotations) == 1:
         # Handle single annotation
-        yield from _read_single_annotation(source_file, annotations[0], with_annotation_name, root, allow_newlines)
+        yield from _read_single_annotation(source_file, annotations[0], with_annotation_name, root)
     else:
         # Handle multiple annotations used as one
 
         # Make sure we don't have multiple attributes on the same annotation
-        assert len(annotations) == len(set(split_annotation(ann)[0]
-                                           for ann in annotations)), "Reading multiple attributes on the same " \
-                                                                     "annotation is not allowed."
+        assert len(annotations) == len(
+            {split_annotation(ann)[0] for ann in annotations}
+        ), "Reading multiple attributes on the same annotation is not allowed."
 
         # Get iterators for all annotations
         all_annotations = {split_annotation(ann)[0]: _read_single_annotation(source_file, ann, with_annotation_name,
-                                                                             root, allow_newlines)
+                                                                             root)
                            for ann in annotations}
 
         # We need to read the annotation spans to be able to interleave the values in the correct order
         for _, ann in heapq.merge(*[_read_single_annotation(source_file, split_annotation(ann)[0],
                                                             with_annotation_name=True,
-                                                            root=root, allow_newlines=allow_newlines)
+                                                            root=root)
                                     for ann in annotations]):
             yield next(all_annotations[ann])
 
 
 def read_annotation_attributes(source_file: str, annotations: Union[List[BaseAnnotation], Tuple[BaseAnnotation, ...]],
-                               with_annotation_name: bool = False, allow_newlines: bool = False):
+                               with_annotation_name: bool = False) -> Iterator[tuple]:
     """Yield tuples of multiple attributes on the same annotation."""
     assert isinstance(annotations, (tuple, list)), "'annotations' argument must be tuple or list"
-    assert len(set(split_annotation(annotation)[0] for annotation in annotations)) == 1, "All attributes need to be " \
-                                                                                         "for the same annotation"
-    return zip(*[read_annotation(source_file, annotation, with_annotation_name, allow_newlines)
+    assert len({split_annotation(annotation)[0] for annotation in
+                annotations}) == 1, "All attributes need to be for the same annotation"
+    return zip(*[read_annotation(source_file, annotation, with_annotation_name)
                  for annotation in annotations])
 
 
-def _read_single_annotation(source_file: str, annotation: str, with_annotation_name: bool, root: Path = None,
-                            allow_newlines: bool = False):
+def _read_single_annotation(
+    source_file: str,
+    annotation: str,
+    with_annotation_name: bool,
+    root: Optional[Path] = None
+) -> Iterator[Any]:
     """Read a single annotation file."""
     ann_file = get_annotation_path(source_file, annotation, root)
 
-    with open_annotation_file(ann_file) as f:
-        ctr = 0
-        try:
-            line: str
-            for line in f:
-                value = line.rstrip("\n\r")
-                if not split_annotation(annotation)[1]:  # If this is a span annotation
-                    value = tuple(tuple(map(int, pos.split("."))) for pos in value.split("-"))
-                elif allow_newlines:
-                    # Replace literal "\n" with line break (if we allow "\n" in values)
-                    value = re.sub(r"((?<!\\)(?:\\\\)*)\\n", r"\1\n", value).replace(r"\\", "\\")
-                yield value if not with_annotation_name else (value, annotation)
-                ctr += 1
-        except (gzip.BadGzipFile, OSError, lzma.LZMAError, UnicodeDecodeError) as e:
-            if isinstance(e, OSError) and str(e) != "Invalid data stream":
-                raise e
-            raise_format_error(ann_file)
-    logger.debug("Read %d items: %s%s", ctr, source_file + "/" if source_file else "", annotation)
+    ctr = 0
+    for value in read_annotation_file(ann_file):
+        yield value if not with_annotation_name else (value, annotation)
+        ctr += 1
+    logger.debug("Read %d items: %s%s%s", ctr, source_file, "/" if source_file else "", annotation)
 
 
-def write_data(source_file: Optional[str], name: Union[BaseAnnotation, str], value: str, append: bool = False):
-    """Write arbitrary string data to file in workdir directory."""
+def write_data(source_file: Optional[str], name: Union[BaseAnnotation, str], value: Any) -> None:
+    """Write arbitrary data to file in workdir directory."""
     file_path = get_annotation_path(source_file, name, data=True)
-    os.makedirs(os.path.dirname(file_path), exist_ok=True)
-    mode = "a" if append else "w"
+    file_path.parent.mkdir(parents=True, exist_ok=True)
 
-    with open_annotation_file(file_path, mode) as f:
-        f.write(value)
+    write_annotation_file(file_path, value, is_data=True)
+
     # Update file modification time even if nothing was written
     os.utime(file_path, None)
-    logger.info(
-        "Wrote %d bytes: %s%s",
-        len(value),
-        source_file + "/" if source_file else "",
-        name.name if isinstance(name, BaseAnnotation) else name
-    )
+    if logger.isEnabledFor(logging.INFO):
+        logger.info(
+            "Wrote %d bytes: %s%s%s",
+            file_path.stat().st_size,
+            source_file,
+            "/" if source_file else "",
+            name.name if isinstance(name, BaseAnnotation) else name
+        )
 
 
-def read_data(source_file: Optional[str], name: Union[BaseAnnotation, str]):
-    """Read arbitrary string data from file in workdir directory."""
+def read_data(source_file: Optional[str], name: Union[BaseAnnotation, str]) -> Any:
+    """Read arbitrary data from file in workdir directory."""
     file_path = get_annotation_path(source_file, name, data=True)
+    data = next(read_annotation_file(file_path, is_data=True))
 
-    with open_annotation_file(file_path) as f:
-        try:
-            data = f.read()
-        except (gzip.BadGzipFile, OSError, lzma.LZMAError, UnicodeDecodeError) as e:
-            if isinstance(e, OSError) and str(e) != "Invalid data stream":
-                raise e
-            raise_format_error(file_path)
-
-    logger.debug(
-        "Read %d bytes: %s%s",
-        len(data),
-        source_file + "/" if source_file else "",
-        name.name if isinstance(name, BaseAnnotation) else name
-    )
+    if logger.isEnabledFor(logging.DEBUG):
+        logger.debug(
+            "Read %d bytes: %s%s%s",
+            file_path.stat().st_size,
+            source_file,
+            "/" if source_file else "",
+            name.name if isinstance(name, BaseAnnotation) else name
+        )
     return data
 
 
@@ -270,7 +238,7 @@ def join_annotation(name: str, attribute: Optional[str]) -> str:
     return ELEM_ATTR_DELIM.join((name, attribute)) if attribute else name
 
 
-def get_annotation_path(source_file: Optional[str], annotation: Union[BaseAnnotation, str], root: Path = None,
+def get_annotation_path(source_file: Optional[str], annotation: Union[BaseAnnotation, str], root: Optional[Path] = None,
                         data: bool = False) -> Path:
     """Construct a path to an annotation file given a source filename and annotation."""
     chunk = ""
@@ -296,18 +264,41 @@ def get_annotation_path(source_file: Optional[str], annotation: Union[BaseAnnota
     return path
 
 
-def open_annotation_file(filename, mode="rt", encoding=None, errors=None, newline=None):
-    """Read and write annotation and data files using different kinds of compression."""
-    if mode in "rwxa":
-        # Text mode is the default for open(), whereas gzip, bz2 and lzma uses binary mode.
-        # We adopt text mode as default.
-        mode += "t"
+def write_annotation_file(file_path: Path, value: Any, is_data: bool = False) -> None:
+    """Write annotation data to a file."""
+    chunk_size = 1000
     opener = _compressed_open.get(compression, open)
-    return opener(filename, mode=mode, encoding=encoding, errors=errors, newline=newline)
+    with opener(file_path, mode="wb") as f:
+        if is_data:
+            pickle.dump(value, f, protocol=pickle.HIGHEST_PROTOCOL)
+        else:
+            for i in range(0, len(value), chunk_size):
+                pickle.dump(value[i:i + chunk_size], f, protocol=pickle.HIGHEST_PROTOCOL)
 
 
-def raise_format_error(file_path):
-    """Raise a SparvErrorMessage about workdir files having the wrong format."""
-    raise SparvErrorMessage(f"Compression of workdir files is set to '{compression}', but '{file_path}' is in another "
-                            "format. Use the configuration key 'sparv.compression' to set the correct compression or "
-                            "use 'sparv clean' to start over with a clean workdir.")
+def read_annotation_file(file_path: Path, is_data: bool = False) -> Iterator:
+    """Return an iterator for reading an annotation file."""
+    opener = _compressed_open.get(compression, open)
+    with opener(file_path, mode="rb") as f:
+        try:
+            if is_data:
+                yield pickle.load(f)
+                return
+            else:
+                try:
+                    while True:
+                        yield from pickle.load(f)
+                except EOFError:
+                    return
+        except pickle.UnpicklingError:
+            raise SparvErrorMessage(
+                "The workdir files for this corpus could not be read. They were probably created using an older "
+                "version of Sparv. Run 'sparv clean' to start over with a clean workdir."
+            ) from None
+        except (gzip.BadGzipFile, OSError, lzma.LZMAError, UnicodeDecodeError) as e:
+            if isinstance(e, OSError) and str(e) != "Invalid data stream":
+                raise e
+            raise SparvErrorMessage(
+                f"Compression of workdir files is set to '{compression}', but '{file_path}' is in another "
+                "format. Use the configuration key 'sparv.compression' to set the correct compression or "
+                "use 'sparv clean' to start over with a clean workdir.") from None